Add provider trait for HasNonceGuard
