+++ crates/cosmos/cosmos-client-components/src/impls/types/transaction.rs
@@ -1,6 +1,7 @@
 use cgp_core::Async;
+use hermes_relayer_components::transaction::traits::nonce::guard::ProvideNonceGuard;
 use hermes_relayer_components::transaction::traits::types::{
-    ProvideFeeType, ProvideNonceType, ProvideSignerType, ProvideTransactionHashType,
+    HasNonceType, ProvideFeeType, ProvideNonceType, ProvideSignerType, ProvideTransactionHashType,
     ProvideTransactionType, ProvideTxResponseType,
 };
 use ibc_proto::cosmos::tx::v1beta1::{Fee, TxRaw};
@@ -10,6 +11,8 @@ use ibc_relayer::keyring::Secp256k1KeyPair;
 use prost::Message as _;
 use tendermint::hash::Hash;
 use tendermint_rpc::endpoint::tx::Response as TxResponse;
+
+use crate::types::nonce_guard::NonceGuard;
 
 pub struct ProvideCosmosTransactionTypes;
 
@@ -63,4 +66,11 @@ where
     Chain: Async,
 {
     type TxResponse = TxResponse;
+}
+
+impl<Chain> ProvideNonceGuard<Chain> for ProvideCosmosTransactionTypes
+where
+    Chain: HasNonceType<Nonce = Account>,
+{
+    type NonceGuard<'a> = NonceGuard<'a>;
 }
