+++ crates/cosmos/cosmos-client-components/src/types/mod.rs
@@ -2,6 +2,7 @@ pub mod channel;
 pub mod connection;
 pub mod events;
 pub mod messages;
+pub mod nonce_guard;
 pub mod packet_lock;
 pub mod payloads;
 pub mod tendermint;
