+++ crates/cosmos/cosmos-relayer/src/impls/transaction/components.rs
@@ -24,6 +24,7 @@ use hermes_relayer_components::transaction::traits::components::tx_fee_estimater
 use hermes_relayer_components::transaction::traits::components::tx_response_querier::TxResponseQuerierComponent;
 use hermes_relayer_components::transaction::traits::components::tx_submitter::TxSubmitterComponent;
 use hermes_relayer_components::transaction::traits::event::TxResponseAsEventsParserComponent;
+use hermes_relayer_components::transaction::traits::nonce::guard::NonceGuardComponent;
 use hermes_relayer_components::transaction::traits::types::{
     FeeTypeComponent, NonceTypeComponent, SignerTypeComponent, TransactionHashTypeComponent,
     TransactionTypeComponent, TxResponseTypeComponent,
@@ -72,6 +73,7 @@ delegate_components! {
             EventTypeComponent,
             SignerTypeComponent,
             NonceTypeComponent,
+            NonceGuardComponent,
             TransactionTypeComponent,
             TransactionHashTypeComponent,
             FeeTypeComponent,
