+++ crates/cosmos/cosmos-relayer/src/impls/transaction/mod.rs
@@ -1,4 +1,3 @@
 pub mod components;
 pub mod fields;
 pub mod log;
-pub mod types;
